Add candles snapshot test; rename Response (without sequence field) to MarketDataResponse

+++ src/wss/v2/market_data_messages.rs
@@ -264,7 +264,7 @@ pub struct Ohlc {
     pub low: Decimal,
     pub close: Decimal,
     pub vwap: Decimal,
-    pub trades: Decimal,
+    pub trades: i64,
     pub volume: Decimal,
     pub interval_begin: String,
     pub interval: i32,
